Update navigation scripts on Other Solutions screen
Fix problem with blue active timer record square failing to move
because $$otherApps = 1 and the scripts fail to turn it off pause
turned on when opening Other Solutions menu.  It turned out that in
addition to changing all buttons scripts to erase $$otherApps variable
( = “”), I also had to update the names of the Budget and Heart
Whisperer solutions in these scripts and in the scripts in those
solutions that set the main window name to the name of the solution.

Total time for fix plus updating GitHub with changes = 1:46