These zip commits are being done now to make a how to movie
I’m not ready yet to put them on GitHub, because I need to finish the
movie explaining how to use them.  So all this adding and deleting of
zip files is part of making this movie, which I make mistakes doing and
so have to repeat this process of adding and then deleting again.